First Release

+++ project/models.py
@@ -3,7 +3,7 @@ from django.db import models
 # Create your models here.
 class Persona(models.Model):
     document_type = models.CharField(max_length=20)
-    document = models.IntegerField()
+    document = models.BigIntegerField()
     name = models.CharField(max_length=20)
     lastname = models.CharField(max_length=20)
     email = models.EmailField()
